Fix issue with FIF_CURRENT_DIR and cd cmd
When sourcing the plugin manually, and you use things like dirstack etc
the cd will push the dir to the stack which is a problem. I have no
issues just removing the cd and pwd cmd, but compat issues may possibly
arise.

Not sure why this wasn't an issue using a plugin manager

Refs:
- https://stackoverflow.com/questions/59895/how-do-i-get-the-directory-where-a-bash-script-is-located-from-within-the-script